data/requirements.txt
autopep8==1.5.7
greenlet==1.1.0
numpy==1.21.0
pandas==1.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
